backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "villa-ops-ai"
version = "0.1.0"
description = "AI-powered operations assistant for villa/hotel property managers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Config
    "pydantic-settings>=2.7.0",
    # Cache
    "redis[hiredis]>=5.2.0",
    # Auth
    "authlib>=1.4.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    # Validation
    "email-validator>=2.1.0",
    "itsdangerous>=2.1.0",
    # HTTP client
    "httpx>=0.28.0",
    # AI / Agent
    "langgraph>=0.2.0",
    "litellm>=1.55.0",
    "mcp>=1.0.0",
    "langchain-litellm>=0.1.0",
    # Payments
    "stripe>=11.0.0",
    # Utilities
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    # Linting & formatting
    "ruff>=0.8.0",
    # Type checking
    "mypy>=1.13.0",
    # SQLAlchemy stubs
    "sqlalchemy-stubs>=0.4",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "authlib.*",
    "jose.*",
    "litellm.*",
    "langgraph.*",
    "langchain_litellm.*",
    "langchain_core.*",
    "mcp.*",
    "sse_starlette.*",
]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.coverage.run]
source = ["app"]
omit = [
    "app/agent/test_agent.py",
    "app/mcp/test_mcp_client.py",
    "app/billing/scripts/*",
    "app/mcp/server.py",
]
concurrency = ["greenlet"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "if TYPE_CHECKING",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
filterwarnings = [
    "ignore::DeprecationWarning",
]
